se adiciona enlaces de retiros y depositos a todas las respuestas se adiciona u nnuevo endpoint de retiro de dinero y deposito de dinero , se adiciona una exepcion nueva para cuando se trate de eliminar una cuenta no existente se adiciona la entidad monto para poder editar el monto de la cuenta mediante un RequestBody se utiliza el query y la notacion modifying en la interface de CuentaRepository para modifcar el valor del monto de la cuenta

+++ src/main/java/com/api/hateoas/service/CuentaService.java
@@ -1,6 +1,7 @@
 package com.api.hateoas.service;
 
 
+import com.api.hateoas.Exception.CuentasNotFoundException;
 import com.api.hateoas.model.Cuenta;
 import com.api.hateoas.repository.CuentaRepository;
 import jakarta.transaction.Transactional;
@@ -29,9 +30,23 @@ public class CuentaService {
     public Cuenta save(Cuenta cuenta){
         return cuentaRepository.save(cuenta);
     }
-
-    public void delete(Integer id ){
+// el metodo eliminar retorna un not found averiguar porque no lanza la exepcion
+    public void delete(Integer id ) throws CuentasNotFoundException {
+        if (!cuentaRepository.existsById(id)){
+            throw new CuentasNotFoundException("Cuenta no encontrada con el ID: " + id);
+        }
         cuentaRepository.deleteById(id);
+    }
+
+
+    public Cuenta depositar(float monto, Integer id){
+        cuentaRepository.actualizarMonto(monto,id);
+        return cuentaRepository.findById(id).get();
+    }
+
+    public Cuenta retirar(float monto, Integer id){
+        cuentaRepository.actualizarMonto(-monto,id);
+        return cuentaRepository.findById(id).get();
     }
 
 }
